refactor all draft order functionality to be in it’s own class and feature gate it.
