+++ src/Domain/Bootstrap.php
@@ -22,6 +22,7 @@ use Automattic\WooCommerce\Blocks\Payments\Integrations\Cheque;
 use Automattic\WooCommerce\Blocks\Payments\Integrations\PayPal;
 use Automattic\WooCommerce\Blocks\Payments\Integrations\BankTransfer;
 use Automattic\WooCommerce\Blocks\Payments\Integrations\CashOnDelivery;
+use Automattic\WooCommerce\Blocks\Domain\Services\DraftOrders;
 
 /**
  * Takes care of bootstrapping the plugin.
@@ -79,7 +80,7 @@ class Bootstrap {
 			$this->container->get( Installer::class );
 			BlockAssets::init();
 		}
-
+		$this->container->get( DraftOrders::class )->init();
 		$this->container->get( PaymentsApi::class );
 		$this->container->get( RestApi::class );
 		Library::init();
@@ -190,6 +191,12 @@ class Bootstrap {
 			Installer::class,
 			function ( Container $container ) {
 				return new Installer();
+			}
+		);
+		$this->container->register(
+			DraftOrders::class,
+			function( Container $container ) {
+				return new DraftOrders( $container->get( Package::class ) );
 			}
 		);
 	}
