+++ src/Library.php
@@ -22,10 +22,6 @@ class Library {
 	public static function init() {
 		add_action( 'init', array( __CLASS__, 'register_blocks' ) );
 		add_action( 'init', array( __CLASS__, 'define_tables' ) );
-		add_filter( 'wc_order_statuses', array( __CLASS__, 'register_draft_order_status' ) );
-		add_filter( 'woocommerce_register_shop_order_post_statuses', array( __CLASS__, 'register_draft_order_post_status' ) );
-		add_filter( 'woocommerce_valid_order_statuses_for_payment', array( __CLASS__, 'append_draft_order_post_status' ) );
-		add_action( 'woocommerce_cleanup_draft_orders', array( __CLASS__, 'delete_expired_draft_orders' ) );
 	}
 
 	/**
@@ -112,78 +108,6 @@ class Library {
 		foreach ( $atomic_blocks as $atomic_block ) {
 			$instance = new \Automattic\WooCommerce\Blocks\BlockTypes\AtomicBlock( $atomic_block );
 			$instance->register_block_type();
-		}
-	}
-
-	/**
-	 * Register custom order status for orders created via the API during checkout.
-	 *
-	 * Draft order status is used before payment is attempted, during checkout, when a cart is converted to an order.
-	 *
-	 * @param array $statuses Array of statuses.
-	 * @return array
-	 */
-	public static function register_draft_order_status( array $statuses ) {
-		$statuses['wc-checkout-draft'] = _x( 'Draft', 'Order status', 'woo-gutenberg-products-block' );
-		return $statuses;
-	}
-
-	/**
-	 * Register custom order post status for orders created via the API during checkout.
-	 *
-	 * @param array $statuses Array of statuses.
-	 * @return array
-	 */
-	public static function register_draft_order_post_status( array $statuses ) {
-		$statuses['wc-checkout-draft'] = [
-			'label'                     => _x( 'Draft', 'Order status', 'woo-gutenberg-products-block' ),
-			'public'                    => false,
-			'exclude_from_search'       => false,
-			'show_in_admin_all_list'    => false,
-			'show_in_admin_status_list' => true,
-			/* translators: %s: number of orders */
-			'label_count'               => _n_noop( 'Drafts <span class="count">(%s)</span>', 'Drafts <span class="count">(%s)</span>', 'woo-gutenberg-products-block' ),
-		];
-		return $statuses;
-	}
-
-	/**
-	 * Append draft status to a list of statuses.
-	 *
-	 * @param array $statuses Array of statuses.
-	 * @return array
-	 */
-	public static function append_draft_order_post_status( $statuses ) {
-		$statuses[] = 'checkout-draft';
-		return $statuses;
-	}
-
-	/**
-	 * Delete draft orders older than a day in batches of 20.
-	 *
-	 * Ran on a daily cron schedule.
-	 */
-	public static function delete_expired_draft_orders() {
-		$count      = 0;
-		$batch_size = 20;
-		$orders     = wc_get_orders(
-			[
-				'date_modified' => '<=' . strtotime( '-1 DAY' ),
-				'limit'         => $batch_size,
-				'status'        => 'wc-checkout-draft',
-				'type'          => 'shop_order',
-			]
-		);
-
-		if ( $orders ) {
-			foreach ( $orders as $order ) {
-				$order->delete( true );
-				$count ++;
-			}
-		}
-
-		if ( $batch_size === $count && function_exists( 'as_enqueue_async_action' ) ) {
-			as_enqueue_async_action( 'woocommerce_cleanup_draft_orders' );
 		}
 	}
 }
